# backend/openapi.yaml
openapi: 3.0.3
info:
  description: API para Sistema de Gestión de Condominios
  title: Casitas Teto API
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /gastos-comunes:
    get:
      parameters:
      - description: ID del condominio
        in: query
        name: condominioId
        required: false
        schema:
          type: integer
      - in: query
        name: residenteId
        schema:
          type: integer
      - in: query
        name: mes
        schema:
          maximum: 12
          minimum: 1
          type: integer
      - in: query
        name: anio
        schema:
          type: integer
      - in: query
        name: estado
        schema:
          enum:
          - PENDIENTE
          - PAGADO
          - VENCIDO
          - MOROSO
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_gastos_comunes_get_200_response"
          description: Lista de gastos comunes
      summary: Listar gastos comunes
      tags:
      - Gastos Comunes
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GastoComunInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GastoComun"
          description: Gasto común creado
      summary: Crear gasto común
      tags:
      - Gastos Comunes
  /gastos-comunes/{gastoComunId}:
    get:
      parameters:
      - description: ID del gasto común
        in: path
        name: gastoComunId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GastoComun"
          description: Detalles del gasto común
      summary: Obtener gasto común
      tags:
      - Gastos Comunes
    put:
      parameters:
      - description: ID del gasto común
        in: path
        name: gastoComunId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GastoComunInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GastoComun"
          description: Gasto común actualizado
      summary: Actualizar gasto común
      tags:
      - Gastos Comunes
  /gastos-comunes/{gastoComunId}/observaciones:
    post:
      parameters:
      - description: ID del gasto común
        in: path
        name: gastoComunId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_gastos_comunes__gastoComunId__observaciones_post_request"
        required: true
      responses:
        "201":
          description: Observación agregada
      summary: Agregar observación al gasto común
      tags:
      - Gastos Comunes
  /gastos-comunes/{gastoComunId}/notificar:
    post:
      parameters:
      - description: ID del gasto común
        in: path
        name: gastoComunId
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: Notificación enviada
      summary: Enviar notificación de gasto común por correo
      tags:
      - Gastos Comunes
...
  /espacios-comunes:
    get:
      parameters:
      - description: ID del condominio
        in: query
        name: condominioId
        required: false
        schema:
          type: integer
      - in: query
        name: tipo
        schema:
          enum:
          - ESTACIONAMIENTO
          - QUINCHO
          - MULTICANCHA
          - SALA_EVENTOS
          type: string
      - in: query
        name: disponible
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/EspacioComun"
                type: array
          description: Lista de espacios comunes
      summary: Listar espacios comunes
      tags:
      - Espacios Comunes
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EspacioComunInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EspacioComun"
          description: Espacio común creado
      summary: Crear espacio común
      tags:
      - Espacios Comunes
  /espacios-comunes/{espacioComunId}:
    get:
      parameters:
      - description: ID del espacio común
        in: path
        name: espacioComunId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EspacioComun"
          description: Detalles del espacio común
      summary: Obtener espacio común
      tags:
      - Espacios Comunes
  /reservas:
    get:
      parameters:
      - description: ID del condominio
        in: query
        name: condominioId
        required: false
        schema:
          type: integer
      - in: query
        name: espacioComunId
        schema:
          type: integer
      - in: query
        name: residenteId
        schema:
          type: integer
      - in: query
        name: fechaDesde
        schema:
          format: date
          type: string
      - in: query
        name: fechaHasta
        schema:
          format: date
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_reservas_get_200_response"
          description: Lista de reservas
      summary: Listar reservas
      tags:
      - Espacios Comunes
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservasInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reserva"
          description: Reserva creada
        "409":
          description: Espacio no disponible
      summary: Crear reserva
      tags:
      - Espacios Comunes
  /reservas/{reservaId}:
    delete:
      parameters:
      - description: ID de la reserva
        in: path
        name: reservaId
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: Reserva cancelada
      summary: Cancelar reserva
      tags:
      - Espacios Comunes
    get:
      parameters:
      - description: ID de la reserva
        in: path
        name: reservaId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reserva"
          description: Detalles de la reserva
      summary: Obtener reserva
      tags:
      - Espacios Comunes
...
components:
...
  schemas:
...
    GastoComun:
      example:
        multas: 3.616076749251911
        estado: PENDIENTE
        fechaVencimiento: 2000-01-23
        residenteId: 6
        fechaEmision: 2000-01-23
        fechaPago: 2000-01-23T04:56:07.000+00:00
        cuotaMantencion: 7.061401241503109
        servicios: 9.301444243932576
        observaciones:
        - texto: texto
          fecha: 2000-01-23T04:56:07.000+00:00
          usuarioId: 4
        - texto: texto
          fecha: 2000-01-23T04:56:07.000+00:00
          usuarioId: 4
        montoBase: 2.3021358869347655
        condominioId: 1
        mes: 5
        id: 0
        año: 5
        montoTotal: 2.027123023002322
      properties:
        id:
          title: id
          type: integer
        residenteId:
          title: residenteId
          type: integer
        condominioId:
          title: condominioId
          type: integer
        mes:
          title: mes
          type: integer
        año:
          title: año
          type: integer
        montoBase:
          format: decimal
          title: montoBase
          type: number
        cuotaMantencion:
          format: decimal
          title: cuotaMantencion
          type: number
        servicios:
          format: decimal
          title: servicios
          type: number
        multas:
          format: decimal
          title: multas
          type: number
        montoTotal:
          format: decimal
          title: montoTotal
          type: number
        estado:
          enum:
          - PENDIENTE
          - PAGADO
          - VENCIDO
          - MOROSO
          title: estado
          type: string
        fechaEmision:
          format: date
          title: fechaEmision
          type: string
        fechaVencimiento:
          format: date
          title: fechaVencimiento
          type: string
        fechaPago:
          format: date-time
          nullable: true
          title: fechaPago
          type: string
        observaciones:
          items:
            $ref: "#/components/schemas/GastoComun_observaciones_inner"
          title: observaciones
          type: array
      title: GastoComun
      type: object
    GastoComunInput:
      example:
        servicios: 0.23021358869347652
        fechaVencimiento: 2000-01-23
        residenteId: 0
        montoBase: 0.5637376656633328
        condominioId: 6
        mes: 2
        año: 5
      properties:
        residenteId:
          title: residenteId
          type: integer
        condominioId:
          title: condominioId
          type: integer
        mes:
          maximum: 12
          minimum: 1
          title: mes
          type: integer
        año:
          title: año
          type: integer
        montoBase:
          format: decimal
          minimum: 0
          title: montoBase
          type: number
        servicios:
          format: decimal
          minimum: 0
          title: servicios
          type: number
        fechaVencimiento:
          format: date
          title: fechaVencimiento
          type: string
      required:
      - año
      - condominioId
      - mes
      - montoBase
      - residenteId
      - servicios
      title: GastoComunInput
      type: object
...
    EspacioComun:
      example:
        descripcion: descripcion
        tipo: ESTACIONAMIENTO
        costoPorHora: 5.962133916683182
        requierePago: true
        condominioId: 6
        id: 0
        nombre: nombre
        capacidad: 1
        activo: true
      properties:
        id:
          title: id
          type: integer
        condominioId:
          title: condominioId
          type: integer
        nombre:
          title: nombre
          type: string
        tipo:
          enum:
          - ESTACIONAMIENTO
          - QUINCHO
          - MULTICANCHA
          - SALA_EVENTOS
          title: tipo
          type: string
        capacidad:
          title: capacidad
          type: integer
        costoPorHora:
          format: decimal
          nullable: true
          title: costoPorHora
          type: number
        descripcion:
          title: descripcion
          type: string
        activo:
          title: activo
          type: boolean
        requierePago:
          title: requierePago
          type: boolean
      title: EspacioComun
      type: object
    EspacioComunInput:
      example:
        descripcion: descripcion
        tipo: ESTACIONAMIENTO
        costoPorHora: 0.14658129805029452
        requierePago: true
        condominioId: 0
        nombre: nombre
        capacidad: 1
      properties:
        condominioId:
          title: condominioId
          type: integer
        nombre:
          title: nombre
          type: string
        tipo:
          enum:
          - ESTACIONAMIENTO
          - QUINCHO
          - MULTICANCHA
          - SALA_EVENTOS
          title: tipo
          type: string
        capacidad:
          minimum: 1
          title: capacidad
          type: integer
        costoPorHora:
          format: decimal
          minimum: 0
          nullable: true
          title: costoPorHora
          type: number
        descripcion:
          title: descripcion
          type: string
        requierePago:
          title: requierePago
          type: boolean
      required:
      - condominioId
      - nombre
      - tipo
      title: EspacioComunInput
      type: object
    Reserva:
      example:
        horaFin: horaFin
        espacioComunId: 6
        estado: PENDIENTE_PAGO
        pagoId: 5
        residenteId: 1
        montoPago: 5.962133916683182
        observaciones: observaciones
        fechaCreacion: 2000-01-23T04:56:07.000+00:00
        id: 0
        horaInicio: horaInicio
        fechaReserva: 2000-01-23
      properties:
        id:
          title: id
          type: integer
        espacioComunId:
          title: espacioComunId
          type: integer
        residenteId:
          title: residenteId
          type: integer
        fechaReserva:
          format: date
          title: fechaReserva
          type: string
        horaInicio:
          format: time
          title: horaInicio
          type: string
        horaFin:
          format: time
          title: horaFin
          type: string
        estado:
          enum:
          - PENDIENTE_PAGO
          - CONFIRMADA
          - CANCELADA
          - COMPLETADA
          title: estado
          type: string
        montoPago:
          format: decimal
          nullable: true
          title: montoPago
          type: number
        pagoId:
          nullable: true
          title: pagoId
          type: integer
        observaciones:
          title: observaciones
          type: string
        fechaCreacion:
          format: date-time
          title: fechaCreacion
          type: string
      title: Reserva
      type: object
...
    ReservasInput:
      example:
        horaFin: horaFin
        espacioComunId: 0
        residenteId: 6
        observaciones: observaciones
        horaInicio: horaInicio
        fechaReserva: 2000-01-23
      properties:
        espacioComunId:
          title: espacioComunId
          type: integer
        residenteId:
          title: residenteId
          type: integer
        fechaReserva:
          format: date
          title: fechaReserva
          type: string
        horaInicio:
          format: time
          title: horaInicio
          type: string
        horaFin:
          format: time
          title: horaFin
          type: string
        observaciones:
          title: observaciones
          type: string
      required:
      - espacioComunId
      - fechaReserva
      - horaFin
      - horaInicio
      - residenteId
      title: ReservasInput
      type: object
...
    _gastos_comunes_get_200_response:
      example:
        data:
        - multas: 3.616076749251911
          estado: PENDIENTE
          fechaVencimiento: 2000-01-23
          residenteId: 6
          fechaEmision: 2000-01-23
          fechaPago: 2000-01-23T04:56:07.000+00:00
          cuotaMantencion: 7.061401241503109
          servicios: 9.301444243932576
          observaciones:
          - texto: texto
            fecha: 2000-01-23T04:56:07.000+00:00
            usuarioId: 4
          - texto: texto
            fecha: 2000-01-23T04:56:07.000+00:00
            usuarioId: 4
          montoBase: 2.3021358869347655
          condominioId: 1
          mes: 5
          id: 0
          año: 5
          montoTotal: 2.027123023002322
        - multas: 3.616076749251911
          estado: PENDIENTE
          fechaVencimiento: 2000-01-23
          residenteId: 6
          fechaEmision: 2000-01-23
          fechaPago: 2000-01-23T04:56:07.000+00:00
          cuotaMantencion: 7.061401241503109
          servicios: 9.301444243932576
          observaciones:
          - texto: texto
            fecha: 2000-01-23T04:56:07.000+00:00
            usuarioId: 4
          - texto: texto
            fecha: 2000-01-23T04:56:07.000+00:00
            usuarioId: 4
          montoBase: 2.3021358869347655
          condominioId: 1
          mes: 5
          id: 0
          año: 5
          montoTotal: 2.027123023002322
      properties:
        data:
          items:
            $ref: "#/components/schemas/GastoComun"
          title: data
          type: array
      title: _gastos_comunes_get_200_response
      type: object
    _gastos_comunes__gastoComunId__observaciones_post_request:
      properties:
        observacion:
          title: observacion
          type: string
      required:
      - observacion
      title: _gastos_comunes__gastoComunId__observaciones_post_request
      type: object
...
    _reservas_get_200_response:
      example:
        data:
        - horaFin: horaFin
          espacioComunId: 6
          estado: PENDIENTE_PAGO
          pagoId: 5
          residenteId: 1
          montoPago: 5.962133916683182
          observaciones: observaciones
          fechaCreacion: 2000-01-23T04:56:07.000+00:00
          id: 0
          horaInicio: horaInicio
          fechaReserva: 2000-01-23
        - horaFin: horaFin
          espacioComunId: 6
          estado: PENDIENTE_PAGO
          pagoId: 5
          residenteId: 1
          montoPago: 5.962133916683182
          observaciones: observaciones
          fechaCreacion: 2000-01-23T04:56:07.000+00:00
          id: 0
          horaInicio: horaInicio
          fechaReserva: 2000-01-23
      properties:
        data:
          items:
            $ref: "#/components/schemas/Reserva"
          title: data
          type: array
      title: _reservas_get_200_response
      type: object
...
    GastoComun_observaciones_inner:
      example:
        texto: texto
        fecha: 2000-01-23T04:56:07.000+00:00
        usuarioId: 4
      properties:
        fecha:
          format: date-time
          title: fecha
          type: string
        texto:
          title: texto
          type: string
        usuarioId:
          title: usuarioId
          type: integer
      title: GastoComun_observaciones_inner
      type: object